+++ next/src/components/articulation/overview/ArticulationOverview.test.tsx
@@ -15,7 +15,7 @@ describe('ArticulationOverview', () => {
       />,
     );
 
-    expect(getByText('Prénom')).toBeInTheDocument();
+    expect(getByText('First Name')).toBeInTheDocument();
     expect(getByText('prenom formula')).toBeInTheDocument();
   });
 });